tweaks from Nick's code review
