+++ pwiz_tools/Skyline/Util/Util.cs
@@ -1951,111 +1951,6 @@ namespace pwiz.Skyline.Util
             return d.HasValue ? d.Value.ToString(LocalizationHelper.CurrentCulture) : String.Empty;
         }
 
-        /// <summary>
-        /// Compare two strings (assumed to be multi-line values separated by newlines), ignoring
-        /// differences that are due to per-run generated values such as GUIDs or timestamps, and
-        /// optionally allowing for small differences in tab separated columnar numerical values
-        /// </summary>
-        /// <param name="expected">expected value</param>
-        /// <param name="actual">actual value</param>
-        /// <param name="columnTolerances">optional dictionary of zero-based column indexes and allowable difference tolerances</param>
-        /// <returns>Empty string if no meaningful differences found, or a string describing the difference</returns>
-        public static string DiffIgnoringTimeStampsAndGUIDs(string expected, string actual, Dictionary<int, double> columnTolerances = null)
-        {
-            string GetEarlyEndingMessage(string name, int count, string lineEqualLast, string lineNext, TextReader reader)
-            {
-                int linesRemaining = 0;
-                while (reader.ReadLine() != null)
-                    linesRemaining++;
-
-                return string.Format(@"{0} stops at line {1}:\r\n{2}\r\n>\r\n+ {3}\r\n{4} more lines",
-                    name, count, lineEqualLast, lineNext, linesRemaining);
-            }
-            using (StringReader readerExpected = new StringReader(expected))
-            using (StringReader readerActual = new StringReader(actual))
-            {
-                int count = 1;
-                string lineEqualLast = string.Empty;
-                while (true)
-                {
-                    string lineExpected = readerExpected.ReadLine();
-                    string lineActual = readerActual.ReadLine();
-                    if (lineExpected == null && lineActual == null)
-                    {
-                        return string.Empty;  // We're done
-                    }
-                    if (lineExpected == null)
-                    {
-                        return GetEarlyEndingMessage(@"Expected", count - 1, lineEqualLast, lineActual, readerActual);
-                    }
-                    if (lineActual == null)
-                    {
-                        return GetEarlyEndingMessage(@"Actual", count - 1, lineEqualLast, lineExpected, readerExpected);
-                    }
-                    if (lineExpected != lineActual)
-                    {
-                        // If only difference appears to be a generated GUID, let it pass
-                        var regexLSID = new Regex(@"(.*)\:[0123456789abcdef]*-[0123456789abcdef]*-[0123456789abcdef]*-[0123456789abcdef]*-[0123456789abcdef]*\:(.*)");
-                        var matchTarget = regexLSID.Match(lineExpected);
-                        var matchActual = regexLSID.Match(lineActual);
-                        if (matchTarget.Success && matchActual.Success
-                                                && Equals(matchTarget.Groups[1].ToString(), matchActual.Groups[1].ToString())
-                                                && Equals(matchTarget.Groups[2].ToString(), matchActual.Groups[2].ToString()))
-                        {
-                            continue;
-                        }
-
-                        // If only difference appears to be a generated ISO timestamp, let it pass
-                        // e.g. 2020-07-10T10:40:03Z or 2020-07-10T10:40:03-07:00 etc
-                        var regexTimestamp = new Regex(@"(.*"")\d\d\d\d\-\d\d\-\d\dT\d\d\:\d\d\:\d\d(?:Z|(?:[\-\+]\d\d\:\d\d))("".*)");
-                        matchTarget = regexTimestamp.Match(lineExpected);
-                        matchActual = regexTimestamp.Match(lineActual);
-                        if (matchTarget.Success && matchActual.Success
-                                                && Equals(matchTarget.Groups[1].ToString(), matchActual.Groups[1].ToString())
-                                                && Equals(matchTarget.Groups[2].ToString(), matchActual.Groups[2].ToString()))
-                        {
-                            continue;
-                        }
-
-                        bool failed = true;
-                        if (columnTolerances != null)
-                        {
-                            // ReSharper disable PossibleNullReferenceException
-                            var colsActual = lineActual.Split('\t');
-                            var colsTarget = lineExpected.Split('\t');
-                            // ReSharper restore PossibleNullReferenceException
-                            if (colsTarget.Length == colsActual.Length)
-                            {
-                                failed = false; // May yet be saved by tolerance check
-                                for (var c = 0; c < colsActual.Length; c++)
-                                {
-                                    if (colsActual[c] != colsTarget[c])
-                                    {
-                                        double valActual, valTarget;
-                                        if (!columnTolerances.ContainsKey(c) ||
-                                            !(double.TryParse(colsActual[c], out valActual) &&
-                                              double.TryParse(colsTarget[c], out valTarget)) ||
-                                            (Math.Abs(valActual - valTarget) > columnTolerances[c] + columnTolerances[c] / 1000)) // Allow for rounding cruft
-                                        {
-                                            failed = true;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        if (failed)
-                        {
-                            return string.Format(@"Diff found at line {0}:\r\n{1}\r\n>\r\n{2}", count, lineExpected, lineActual);
-                        }
-                    }
-
-                    lineEqualLast = lineExpected;
-                    count++;
-                }
-            }
-        }
     }
 
     /// <summary>
